requirements.txt
aioflask==0.4.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
dnspython==2.3.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.95.2
Flask==2.3.2
Flask-Cors==3.0.10
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
greenletio==0.10.1
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.23.1
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
loguru==0.7.0
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.3
openai==0.27.7
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==3.5.1
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pydantic==1.10.8
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
regex==2023.5.5
requests==2.31.0
rich==13.3.5
six==1.16.0
smmap==5.0.0
snakeMQ==1.6
sniffio==1.3.0
stack-data==0.6.2
starlette==0.27.0
streamlit==1.22.0
tenacity==8.2.2
tiktoken==0.4.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.6.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.2
uvicorn==0.22.0
validators==0.20.0
wcwidth==0.2.6
Werkzeug==2.3.4
yarl==1.9.2
zipp==3.15.0
